mapToSyncData was split into mapToDTOs and fromDTO so that every individual character mapping can throw an exception without interrupting the whole stream like before. Adapted the Service to make use of that change in a straightfoward fashion for now. Changed some exception messages, solidified ProfileMapper in general, made profile dtos contain Wrapper fields instead of primitives

+++ src/test/java/com/bebis/BeBiS/profile/ProfileServiceIT.java
@@ -74,9 +74,9 @@ public class ProfileServiceIT extends BaseFullStackTest {
         callService();
 
         // then
-        List<Integer> savedIds = jdbcTemplate.queryForList("SELECT id FROM wow_characters WHERE blizzard_account_id = ?",
-                Integer.class, BLIZZ_ACCOUNT_ID);
-        assertThat(savedIds).doesNotContain((int) oldDTO.wowCharacterId());
+        List<Long> savedIds = jdbcTemplate.queryForList("SELECT id FROM wow_characters WHERE blizzard_account_id = ?",
+                Long.class, BLIZZ_ACCOUNT_ID);
+        assertThat(savedIds).doesNotContain(oldDTO.wowCharacterId());
     }
 
     @Test
